# docs/swagger.yaml
definitions:
...
  models.Login:
    properties:
      code:
        type: string
      password:
        type: string
      username:
        type: string
      uuid:
        type: string
    required:
    - code
    - password
    - username
    - uuid
    type: object
...
info:
  contact: {}
  description: |-
    基于Gin + Vue + Element UI前后端分离的工单系统的接口文档
  license:
    name: MIT
    url: https://github.com/csthenry/ferry-backend/blob/main/LICENSE
  title: ferry API
  version: 1.0.0
paths:
...
  /login:
    post:
      consumes:
      - application/json
      description: 获取token
      parameters:
      - description: Add account
        in: body
        name: username
        required: true
        schema:
          $ref: '#/definitions/models.Login'
          type: object
      responses:
        "200":
          description: '{"code": 200, "expire": "2019-08-07T12:45:48+08:00", "token":
            ".eyJleHAiOjE1NjUxNTMxNDgsImlkIjoiYWRtaW4iLCJvcmlnX2lhdCI6MTU2NTE0OTU0OH0.-zvzHvbg0A"
            }'
          schema:
            type: string
      summary: 登录
  /logout:
    post:
      consumes:
      - application/json
      description: 获取token
      responses:
        "200":
          description: '{"code": 200, "msg": "成功退出系统" }'
          schema:
            type: string
      security:
      - "": []
      summary: 退出登录
...
securityDefinitions:
  Bearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"